Fix bundling execptions

+++ assemblyline/common/bundling.py
@@ -30,19 +30,19 @@ class BundlingException(Exception):
     pass
 
 
-class AlertNotFound(Exception):
+class AlertNotFound(BundlingException):
     pass
 
 
-class SubmissionNotFound(Exception):
+class SubmissionNotFound(BundlingException):
     pass
 
 
-class IncompleteBundle(Exception):
+class IncompleteBundle(BundlingException):
     pass
 
 
-class SubmissionAlreadyExist(Exception):
+class SubmissionAlreadyExist(BundlingException):
     pass
 
 
@@ -223,7 +223,8 @@ def create_bundle(sid, working_dir=WORK_DIR, use_alert=False):
                         pack_stream(ih, oh, {'al': {"type": BUNDLE_TYPE}, 'name': f"{sid}.tgz"})
 
                 return target_file
-
+        except (SubmissionNotFound, AlertNotFound):
+            raise
         except Exception as e:
             raise BundlingException("Could not bundle submission '%s'. [%s: %s]" % (sid, type(e).__name__, str(e)))
         finally:
@@ -235,7 +236,7 @@ def create_bundle(sid, working_dir=WORK_DIR, use_alert=False):
 
 # noinspection PyBroadException,PyProtectedMember
 def import_bundle(path, working_dir=WORK_DIR, min_classification=Classification.UNRESTRICTED, allow_incomplete=False,
-                  rescan_services=None, completed_queue=None, exist_ok=False):
+                  rescan_services=None, completed_queue=None, exist_ok=False, cleanup=True):
     with forge.get_datastore(archive_access=True) as datastore:
         current_working_dir = os.path.join(working_dir, get_random_id())
         res_file = os.path.join(current_working_dir, "results.json")
@@ -353,17 +354,11 @@ def import_bundle(path, working_dir=WORK_DIR, min_classification=Classification.
 
             return submission
         finally:
-            try:
+            if extracted_path != path and os.path.exists(extracted_path):
                 os.remove(extracted_path)
-            except Exception:
-                pass
 
-            try:
+            if cleanup and os.path.exists(path):
                 os.remove(path)
-            except Exception:
-                pass
 
-            try:
+            if os.path.exists(current_working_dir):
                 shutil.rmtree(current_working_dir, ignore_errors=True)
-            except Exception:
-                pass
